[GOHW-2]: переход на grpc

See merge request wgwergwegwer/homework-1!2

+++ checkout/internal/domain/types.go
@@ -8,7 +8,7 @@ type Stock struct {
 type CartItem struct {
 	UserId uint32
 	Sku    uint32
-	Count  uint16
+	Count  uint32
 }
 
 type ProductInfo struct {
